requirements.txt
ipykernel
pandas
pyarrow
fastparquet
nltk
langdetect

pinecone
openai
googletrans

fastapi
fastapi[standard]

python-dotenv